COLAMD and CCOLAMD: split out *_version method into its own file
