+++ CCOLAMD/Source/ccolamd.c
@@ -973,20 +973,6 @@ PRIVATE void print_report
 /* === USER-CALLABLE ROUTINES: ============================================== */
 /* ========================================================================== */
 
-//------------------------------------------------------------------------------
-// ccolamd_version: return CCOLAMD version
-//------------------------------------------------------------------------------
-
-#ifndef DLONG
-void ccolamd_version (int version [3])
-{
-    version [0] = CCOLAMD_MAIN_VERSION ;
-    version [1] = CCOLAMD_SUB_VERSION ;
-    version [2] = CCOLAMD_SUBSUB_VERSION ;
-}
-#endif
-
-
 /* ========================================================================== */
 /* === ccolamd_recommended ================================================== */
 /* ========================================================================== */
