+++ COLAMD/Source/colamd.c
@@ -974,19 +974,6 @@ PRIVATE void debug_structures
 /* === USER-CALLABLE ROUTINES: ============================================== */
 /* ========================================================================== */
 
-//------------------------------------------------------------------------------
-// colamd_version: return COLAMD version
-//------------------------------------------------------------------------------
-
-#ifndef DLONG
-void colamd_version (int version [3])
-{
-    version [0] = COLAMD_MAIN_VERSION ;
-    version [1] = COLAMD_SUB_VERSION ;
-    version [2] = COLAMD_SUBSUB_VERSION ;
-}
-#endif
-
 /* ========================================================================== */
 /* === colamd_recommended =================================================== */
 /* ========================================================================== */
